+++ packages/react-static/src/static/__mocks__/configWithPluginWithOptions.mock.js
@@ -1,4 +1,3 @@
 export default {
-  entry: 'path/to/entry/index.js',
   plugins: [['mock-plugin', { mockOption: 'some-option' }]],
 }