+++ packages/react-static/src/static/__mocks__/defaultConfigDevelopment.mock.js
@@ -5,7 +5,7 @@ export default {
   inlineCss: false,
   outputFileRate: 100,
   extractCssChunks: false,
-  entry: './root/src/index.js',
+  entry: 'index.js',
   paths: {
     ROOT: './root/',
     SRC: './root/src',