+++ packages/react-static/src/static/__mocks__/static.config.mock.js
@@ -1,3 +1,1 @@
-export default {
-  entry: 'path/to/entry/index.js',
-}
+export default {}