+++ packages/react-static/src/static/webpack/webpack.config.dev.js
@@ -8,7 +8,7 @@ import path from 'path'
 import rules from './rules'
 
 export default function({ config }) {
-  const { ROOT, DIST, NODE_MODULES, SRC, HTML_TEMPLATE } = config.paths
+  const { DIST, NODE_MODULES, SRC, HTML_TEMPLATE } = config.paths
 
   process.env.REACT_STATIC_BASE_PATH = config.basePath
   process.env.REACT_STATIC_PUBLIC_PATH = config.publicPath
@@ -32,7 +32,7 @@ export default function({ config }) {
             require.resolve('../../bootstrapPlugins'),
             require.resolve('../../bootstrapTemplates'),
           ]),
-      path.resolve(ROOT, config.entry),
+      config.entry,
     ].filter(Boolean),
     output: {
       filename: '[name].js', // never hash dev code