+++ packages/react-static/templates/typescript/src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Root, Routes, addPrefetchExcludes } from 'react-static'
-import { Link, Router } from 'components/Router'
+import { Link, Router } from '@reach/router'
 import FancyDiv from 'components/FancyDiv'
 import Dynamic from 'containers/Dynamic'
 import './app.css'